cdk/awscdk/ctdc/ctdc/requirements.txt
-e .
aws-cdk.core
aws-cdk.aws_iam
aws-cdk.aws_ec2
aws-cdk.aws_opensearchservice
configparser
boto3
python-dotenv